==> 4_condRandForest.R <==

## Run a conditional random forest 


## Load required packages 
require(raster)
library(party)
library(dplyr)
require(rgdal)


## Set the working directory 
setwd("C:/Users/mmama/Downloads/GIS714")


## Read in lakes data to crop ci rasters 
lakes <- readOGR("data/lakes","updatedValidLakes")
jesup.shp <- subset(lakes, GNIS_NAM_1 == "Jesup, Lake")


## Define a vector of months 
months <- c("JAN","FEB","MAR","APR","MAY","JUN","JUL","AUG","SEP","OCT","NOV","DEC")


## Define a results dataframe 
df.results <- as.data.frame(matrix(ncol = 13, nrow = 5))
colnames(df.results) <- c("VAR", months)
df.results$VAR <- c("SECCHI","DO","TN","TP","TURB")


## Loop through each month 
for(month in months){
  
  
  ## Read in CI data and mask 
  ci <- raster(paste0("data/ci/", month, "_2018.tif"))
  ci.crop <- crop(ci, extent(spTransform(jesup.shp, projection(ci))))
  ci.mask <- mask(ci.crop, spTransform(jesup.shp, projection(ci)))
  ## Read in environmental data 
  secchi <- raster(paste0("results/", month, "_SECCHI.tif"))
  do <- raster(paste0("results/", month, "_DO.tif"))
  tn <- raster(paste0("results/", month, "_TN.tif"))
  tp <- raster(paste0("results/", month, "_TP.tif"))
  turb <- raster(paste0("results/", month, "_TURB.tif"))  
  
  
  ## Create a dataframe of values 
  df <- as.data.frame(matrix(nrow = 1260, ncol = 6))
  colnames(df) <- c("CI","SECCHI","DO","TN","TP","TURB")
  df$CI <- values(ci.mask)
  df$SECCHI <- values(secchi)
  df$DO <- values(do)
  df$TN <- values(tn)
  df$TP <- values(tp)
  df$TURB <- values(turb)
  
  
  ## Create a conditional random forest 
  form <- as.formula("CI ~ SECCHI + DO + TN + TP + TURB")
  crf <- cforest(formula = form, data = df[complete.cases(df),], controls = cforest_unbiased(mtry = 5))
  
  
  ## Assess variable importance and save results 
  crf.imp <- varimp(crf, conditional = T)
  df.results[,which(colnames(df.results) == month)] <- as.numeric(crf.imp)
  
}


## Save results 
write.csv(df.results, "varimp.csv", row.names = F)
